Adding Ming as 'submodule'
When you first checkout this change or freshly clone the forge repo
you must now follow up with

    git submodule init
    git submodule update

This may be more work than just easy_install-ing Ming, or perhaps
using pip to install directly from Ming's repo.  If so, all we
have to do is delete the submodule, so there's no harm in trying the
"Git way".  But it sure does strike me as inconvenient.