FORGE-214 - Add oembed support to our markdown extension
This commit includes oembed support via the following syntaxes:
  [embed#URL]
  [embed#(WIDTH,HEIGHT)URL]
where URL, WIDTH, and HEIGHT are user-provided and the rest are literal.

All oembedded objects are served from an iframe on the host specified in the
.ini file.  For development.ini, for instance, this is oembed.localhost.localdomain:8080.
This is in order to prevent cross-site-scripting attacks.  This commit provides support
for a number of different OEmbed providers; see oembed.ini for more details.

This commit also moves support for WikiWords into a separate Markdown instance
with an extension that handles the CamelCase recognition and auto-linking.