FORGE-140 - Related artifact support
This commit supplies two additional fields on Artifacts: references and
backreferences.  References is a list of artifacts referred to by this
artifact.  Backreferences is a list of artifacts that refer to this
artifact.  These lists should be automatically maintained by the search
reactors, and so should be as up-to-date as the SOLR index.

The method of maintaining the fields is by scanning the text to be indexed
in SOLR for artifact shortlinks.  If a shortlink is found and refers to an
existing artifact, then the ref/backref pair is created.