First pass at a user home page
Email addresses, OpenIDs, and Widgets are only displayed for the
logged-in user.  Projects of which the user is a member are listed for
all.  Now we need to talk about if we're factoring the user neighborhood
in a way where the user app can catch /users/me/admin, and if we can go
to /users/you/ and _not_ be redirected home.