FORGE-253: Add a nose plugin to squelch warnings
You must run

  cd NoWarnings
  python setup.py develop

This will happen automatically as part of setup-common.bash if you
happen to use that script (as the sandboxes do).  Then the nowarnings
plugin becomes available in nosetests, and you can activate it by saying

  nosetests --with-nowarnings

This is useful if unfixable warnings from library code are polluting the
test output.