FORGE-319: add 'select' as a type for tracker custom-fields
Every place where custom fields allow editing, build a select for
custom-fields where field.type is select.

It's easiest to split the options on whitespace; anything more
complicated would require re.split.  For uniformity, split status names
on whitespace, too.  This will require setup-app to be run to get the
built-in status names setup correctly, or else hand edited to drop
existing commas.

The only thing left to do is to hide/show the options field during
tracker admin based on whether the field in question is of type
'select'.  We do that with the new JS show_hide_options() and an
"onchange" handler for the select.