BugFix: handle custom-fields on a new ticket
Originally, the new ticket machinery was not designed to handle
custom-fields.  It assumes every field returned by the form is data to
be set directly on the ticket itself.  Custom-fields, however, are not
part of that schema and must be put in the_ticket.custom_fields dict.

This fix tests each returned key.  Those for custom-fields go into the
dict; others onto the ticket itself.