FORGE-374: remove WMD logo
WMD code is not ammenable to change, so do this the hacky way: on
document.ready, use jQuery to find and remove the link and the logo img
it contains.  The span immediately preceding the link contains the tool
icons and must have its padding-right adjusted to no longer make space
for the logo.  I repeat, this is a hack the equivalent of
screen-scraping.