Hook up paging and sorting for Tracker's search and mass-edit pages
The JavaScript is re-usable in any context that uses the query
parameters 'limit', 'page', 'q', and 'sort' (or subset of these).  It is
tied into the template by "exporting" the same-named variables into
JavaScript; and by labeling the sortable columns with a data-sort
attribute to be used in the 'sort' parameter.  Included CSS adds hover
style for the affected column heads and page buttons.

The actual controller methods are simplified, as they don't need to do
anything beyond validation to the aforementioned params, everything can
be in the **kw dict.  The code actually works by simply tweaking the
query parameters and reloading the page, no AJAX.