Add mod_date to the base Artifact schema
We automate setting this field with a mongometa before_save function,
which comes into play when dirty objects get flushed back into Mongo. We
index it as mod_date_dt. Wiki already uses the timestamp on the
versioned history of its page for this purpose, so no need to change
code there.  Tracker, however, has been using created_date in several
places where mod_date was desired.

Because this is a simple schema addition with an if_missing clause, no
migration is required.  The mod_date for existing objects will be set to
the first time they accessed after the new code is in place.