[#347] feedback and automatic retry for form submission
Automatically applied to any form with class="can-retry".  This patch
adds it to ticket and wiki-page editing.  We can add others as we decide
which ones to include.

The behavior is to (1) display immediate feedback "saving...", then if
the page hasn't reloaded from the form submission within 7 seconds, to
display the error "The server is taking too long to respond.  Retrying
in 30 seconds."; and (3) resubmitting the form after 30 total seconds
have elapsed --- putting you into a loop that is only ended by a
successful server response, or user navigation.