[#431] unify flash, user notifications, ajaxy saving... messages
There is now a single fixed-position pane that holds all
user-notifications.

Added a new JS function, flash(html, kind), to add new notifications to
the pane, e.g., flash('<p>Saving...</p>').  Now use the flash function
from the old form-retry code.

User-notifications are generated directly in the pane, but the webflash
message must be moved there with JS.  It is left visible so that in the
case of no JS, it still shows up, just with slightly different texture
than if it had been moved into the notifications pane.