[#628] make the gravatar genshi function take a User instead of an email
This simplifies most of the calls and allows me to specify a default
that is fetched from the /u/username/user_icon, which will be their icon
if they set one or else the generic silhouette.

For those few cases where we really only had an email address, use the
old function just renamed to email_gravatar.