[#930] Allow to remove inner grid width spec which resets <div> rendering direction.
With this grid_12, <div>s are rendered vertically, w/o - horizontally,
and the latter is needed at some places (like neighborhood project list).
(Actually, original genshi page had even simpler layout, but I traced
noticeable rendition discrepancies down to this class, and work it around with it,
to not create even more complex block overrides).
Of course, this is HACK - nothing in CSS explicitly sets such an order, and
there should be classes which do.