[#942][#688] - site admin and stats page, updates to allow caching of site_style.css
The site admin is accessible at /nf/admin/, and you must be an admin on the 'allura'
project to view it.  Currently it's slim, but is expected to grow.

The site_style.css caching is done by the use of a helper to generate the site_style
hrefs which incorporate a cache-busting parameter computed from the contents of the theme
used to generate the site_style.css file itself.