[#1048] add close-boxes to flash messages
Because the close-box is additional content within the message div,
we need to be more careful.  We can't just replace the entire html
contents anymore the way we did in the #save-message case.  Work
directly with the paragraph object instead.  Increase the padding
within messages so text won't overlap the close-box.