[#1087] promote subject generation from the Thread to the underlying artifact
In the case of tickets, the initial thread's generated subject, '#yyy
discussion', was not actually helpful.  In the case of wiki, the thread
remembered a subject based on the initial name of the page, which would
become wrong if the page is renamed.  So removed the setting of subject
in those cases and made the default thread subject be ''.  Post.subject
now tries to get the thread's subject first, and fall back to the
underlying artifact's email_subject --- which can be kept up-to-date
with name or summary changes.  If a thread really has a subject, for
whatever reason, then that takes precedence.