[#1181] script to clear out ticket thread subjects
Earlier, we improved the subject line for ticket related emails
(e0dbb34).  Before that fix, a ticket's thread used a one-time generated
message of the form '#yyy discussion'.  That subject was stored in the
Thread itself and used as the subject on all emails sent by the ticket.
By promoting subject generation up to the ticket, we could include the
ticket summary, giving, e.g., '#yyy Some problem needs fixing'.  A much
more helpful email subject.

All tickets created after e0dbb34 went live now have no subject stored
in the Thread, and so use Ticket.email_subject to get the helpful
subject.

Unfortunately, existing tickets _don't_ get this behavior, because their
Threads already have a (bad) subject.  When the Thread has a good
subject it will be more specific than the underlying artifact, and so we
want it to 'win'.  We don't want any of the existing ticket Thread
subjects to win, so we need a script to clear them all out.  That's what
this script does.