[#1134] Update Ming and Allura to pymongo 1.9
This commit updates our support to pymongo 1.9.  In doing so, there was a good
bit of refactoring of attachment code to be done since pymongo 1.7-1.9 have an
upgraded version of GridFS.  There was also a minor change in pymongo where
pymongo.bson was renamed to plain old bson.  In order to isolate us (somewhat)
from future changes, all the metadata about files was moved from the foo.files
collection to the (newly-created) foo collection, which maintains a pointer into
the foo.files collection.  There are also several convenience methods added,
including a .serve() method on attachments that correctly sets the WSGI response
parameters for serving up attachments.