[#1198] (re)set some default tools for new projects
The whole point of the new add_project page was to make it easy to add a
default set of tools.  However when we re-wrote it, those defaults lost
their initial checked state.  This commit restores their previous
checked values using setdefault, so that when the widget reports an
error, the actual values will override our defaults.