[#1175] Fix problem that prevents data from propagating back to SFX
Previously, we used a M.ProjectRole.query.get(...) to find the admin role.  This
works fine as long as there's only one role with the given name in a database.
However, with the new database sharding, we can have lots, and everywhere except
for here was correctly updated to access M.ProjectRole.by_name(...), which has
the correct logic.