[#1228] Fix issue in migrate-project-database.py
Previously, we were not correctly updating the project roles when migrating
projects to the sharded environment.  This patch corrects the update() statment,
as well as providing a small fix to the Project.roles property, which previously
tested for the existence of a 'a' property on ProjectRoles, where it should have
tested for 'project_id'.