[#1341] Fix emailed context
So it turns out that the error_message argument is only used by WebError for
displaying a message to the browser (irrelevant to reactors).  The supplemental
information is obtained from a magic local variable __traceback_supplement__
which is mostly undocumented except in docstrings.  For more info, within the
python shell type:

>>> import weberror
>>> help(weberror.collector.ExceptionCollector)