1490 Switch to linked title.
This commit reverts a previous change that made the entire row
clickable. What we found was that making the row clickable
prevented users from right or middle clicking on the linked number.
As a compromise we're linking the ticket title to provide a large
click target while at the same time removing JS fanciness to allow
for all the various types of clicks.

Signed-off-by: Kyle Adams <kyle@geek.net>