[#1516] - Rework middleware, remove dependency on MagicalC and threadlocal Environ
To implement this, there was a lot of reworking of our "global"-ish objects (pylons.c,
pylons.g, ew.widget_context, credentials, etc.) to use Paste's StackedObjectProxies (SOPs).
A SOP is managed by a Registry, and all the SOPs can be cleaned up at once.  In order to make
this work, we needed to use the RegistryMiddleware in 'streaming' mode so that it didn't try
to clean up stuff until the iterator was exhausted (as we tend to *use* pylons.c during rendering,
flushing ORM sessions at the end of the stream, etc.).  All the tests now work locally, will trust
Chuck to find anything else I missed.

Signed-off-by: Rick Copeland <rcopeland@geek.net>