[#1551] Make 'Code'/'code' the default mount_label/mount_point for all SCMs
Add JS to make the three SCM checkboxes on the project start page act
like radio buttons.  Else if we let the user start with multiple, we'd
give them all the same name and that would error out in the middle. Also
took this opportunity to remove the long-unused (since ab1d34e)
starter_mounts method.

Signed-off-by: Wolf <wolf@geek.net>