[#1618] Separate classes for ApiToken and ApiTicket.
Special API key is now ApiTicket, a separate model, as it should be
managed/stored separately from API keys which are now stored as SFX
user preferences. Also, actually store it in a separate collection because
of this, and because it doesn't have unique-per-user constraint, and
because its key is being prefixed with 'tck' to avoid mixing up API keys
and API tickets.

Signed-off-by: Paul Sokolovsky <psokolovsky@geek.net>