Revert "[#3634] Added a check to find and fix svn repos that didn't refresh properly"
This was a bandaid that wasn't very appropriate.  It made it very easy for a refresh to be
initiated when an import was already in progress, which caused data inconsistencies
and errors.  If SVN repos end up 'behind' again, we could consider adding this back in a
safer manner, or better, fix the root cause.

This reverts commit 13288d8858e5c70a70414a1047b7d0bda56bbd1f.