[#6239] Remove RelativeLinkRewriter short-circuit
Not sure why it was there to begin with. Asked Rick and he
didn't remember. In any case, removing it has no effect that
I could see. Tests all pass, and artifact links in emails and
web pages are still generated correctly (absolute links in
emails, and relative links in web pages).

Signed-off-by: Tim Van Steenburgh <tvansteenburgh@gmail.com>