Revert "[#6284] ticket:418 commit revision number in svn commit browser"
This was also showing up on the individual commit page, which was
unnecessary since the info was already on that page.  If we wanted something
just for the view within commit_browser, we could use Allura/allura/templates/repo/commit_basic.html
but it seemed out of place and not too necessary there also.

This reverts commit 67593942258eaff592128d1997b244e192c355be.