[#7029] Fixed test interaction issue on auth page extensions test
Had to move the lookup to runtime, because the AuthController was being
instantiated in another test and stored on a BasetestProjectRootController
instance that was difficult to get at to modify for the test.

Also refactored the additional_urls method to return a mapping, because
that makes more sense.  The order wasn't being preserved, anyway.

Signed-off-by: Cory Johns <cjohns@slashdotmedia.com>