[#7824] remove redirect if proj nbhd name differs from nbhd name
This could potentially cause problems when neighborhood.cache.duration is set.

I believe this redirect harkens back to some early days of Allura where
project shortnames were unique across the system, not per-neighborhood.  So
if a project was accessed via the wrong neighborhood, this would redirect it
to the canonical neighborhood.  But now that's not even possible.