[#7837] ticket:738 Fix repo initialization in tests
The problem with previous configuration is in the following:

1. `with_tool` installs app, which creates Repository model
2. test setup in most places creates *another* Repository model in a context of
   app installed by `with_tool`. This makes two Repositories linked to one app.
3. `app.repo` returns first Repository from the DB for app, which can be any of
   those two, depending on test run

That's why some tests, that depend on `app.repo` have failed occasionally (e.g.
`test_webhook_payload` for both git and svn)

Fixed all the tests I found to avoid this problems in the future. It took me
awhile to figure out what's wrong.