[#7878] ticket:771 Convert args in test app to plain string
WebTest==1.4.0 which we are using does not handle unicode well. For example
methods to send GET/POST/etc request raise exception if url is unicode or some
of the parameters are.

Seems like newer versions (2.0.18) supports it better, but we can't use it,
because it depends on WebOb>=1.2 which conflicts with TurboGears which depends
on WebOb==1.0.8.  That is why manual fix for now.