Fix git merge requests to not update project last_updated when viewed.
On merge requests, the sidebar_menu logic to reorder branches (
branches.remove and branches.insert) were often acting on a ming
instrumented list.  That caused the change to be recorded as a change
to the project.  Now the caching code returns a plain list, never
and instrumented one, so that can't happen.

Not sure why this was apparently only happening on merge request
views and not other repo pages, but a good change nonetheless.